moved Status Tile to separate file
