+++ src/components/schedule/ScheduleTiles.js
@@ -1,6 +1,6 @@
 import './schedule.css';
 import { CARS } from '../../data/cars';
-import { COLORS } from '../../values/colors';
+import { StatusTile } from './StatusTile';
 
 
 // Full Table
@@ -56,46 +56,6 @@ export {ScheduleCarTile, ScheduleTable, ScheduleCarImg };
 function EmptyScheduleTile(props){
     return (
         <td></td>
-    )
-}
-
-// Formatted Tile
-
-const colorMap = {
-    // this first part is used for availability
-    Unavailable: COLORS.Red300,
-    Available: COLORS.Green400,
-    Returned: COLORS.Green600,
-    Rented: COLORS.Blue300,
-    Transfer: COLORS.Purple300,
-    Service: COLORS.Yellow400,
-}
-
-function StatusTile(props) {
-    console.log (props.car.Status)
-    let bgColor;
-    switch (props.car.Status) {
-        case "Rented":
-            bgColor = colorMap.Rented
-            break;
-        
-        case "Returned":
-            bgColor = colorMap.Returned
-            break;
-
-        case "Transfer":
-            bgColor = colorMap.Transfer
-            break;
-
-        case "Unavailable":
-            bgColor = colorMap.Unavailable
-            break;
-
-        default:
-            break;
-    }
-    return (
-        <td style={{backgroundColor:bgColor}}/>
     )
 }
